+++ examples/todo/todo/app.py
@@ -1,4 +1,5 @@
 from flask import Flask
+from flask_rebar import HeaderApiKeyAuthenticator
 from marshmallow import fields, pre_dump, pre_load
 
 from flask_rebar import (
@@ -25,3 +26,22 @@ registry = rebar.create_handler_registry(
     swagger_generator=generator,
     handlers="todo.handlers",
 )
+
+
+def create_app():
+    app = Flask(__name__)
+
+    authenticator = HeaderApiKeyAuthenticator(header="X-MyApp-Key")
+    # The HeaderApiKeyAuthenticator does super simple authentication, designed for
+    # service-to-service authentication inside of a protected network, by looking for a
+    # shared secret in the specified header. Here we define what that shared secret is.
+    authenticator.register_key(key="my-api-key")
+    registry.set_default_authenticator(authenticator=authenticator)
+
+    rebar.init_app(app=app)
+
+    return app
+
+
+if __name__ == "__main__":
+    create_app().run()
